:sparkles: login routes created
The users will be created by the admin via HTTP request, the admin will send to the user his username and his password to use the system. The password will encrypted and the request will return a token that lasts for one day. The route to auth is 'api/auth/login' and needs to receive the name and password on the request body (json)